+++ mempool/src/counters.rs
@@ -4,6 +4,16 @@
 use once_cell::sync::Lazy;
 use prometheus::{IntCounter, IntCounterVec};
 
+/// Counter of pending network events to Mempool
+pub static PENDING_MEMPOOL_NETWORK_EVENTS: Lazy<IntCounterVec> = Lazy::new(|| {
+    register_int_counter_vec!(
+        "mempool_pending_network_events",
+        "Counters(queued,dequeued,dropped) related to pending network notifications to Mempool",
+        &["state"]
+    )
+    .unwrap()
+});
+
 pub static MEMPOOL_SERVICE: Lazy<IntCounterVec> = Lazy::new(|| {
     register_int_counter_vec!(
         "libra_mempool_service",
